requirements.txt
sqlalchemy
psycopg2-binary
openai
requests
beautifulsoup4
python-dotenv
supabase